feat(CubeCamera): new useCubeCamera hook
